+++ src/models/index.ts
@@ -1,5 +1,7 @@
 export type InputKeys = 'words' | 'uppercase' | 'numbers' | 'randomChars' | 'language' | 'passLength' | 'separator'
 
+export type PassLength = string | number
+
 export interface InputValue {
 	value?: string // used in randomchars
 	selected: boolean