+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- Added support for sending a number, not just a string for passLength.
+
+### Fixed
+
+- Tests failing, added proper error throwing exception handling for async functions as per vitest docs
+
 ## [1.0.0](https://github.com/ogdakke/api-sala/releases/1.0.0) - 12-10-2023
 
 ### Breaking